@@ -1,35 +1,32 @@
 {
-	"name": "libri",
+	"name": "libri_frontend",
 	"version": "0.0.1",
-	"private": true,
+	"type": "module",
 	"scripts": {
 		"dev": "vite dev",
 		"build": "vite build",
 		"preview": "vite preview",
-		"lint": "prettier --check . && eslint .",
-		"format": "prettier --write ."
+		"format": "prettier --write .",
+		"lint": "prettier --check . && eslint ."
 	},
 	"devDependencies": {
-		"@sveltejs/adapter-auto": "^3.0.0",
 		"@sveltejs/kit": "^2.0.0",
-		"@sveltejs/vite-plugin-svelte": "^3.0.0",
+		"@sveltejs/vite-plugin-svelte": "^4.0.0",
 		"@types/eslint": "^9.6.0",
 		"autoprefixer": "^10.4.20",
-		"eslint": "^9.0.0",
+		"eslint": "^9.7.0",
 		"eslint-config-prettier": "^9.1.0",
 		"eslint-plugin-svelte": "^2.36.0",
 		"globals": "^15.0.0",
-		"postcss": "^8.4.47",
-		"prettier": "^3.1.1",
-		"prettier-plugin-svelte": "^3.1.2",
-		"svelte": "^4.2.7",
-		"tailwindcss": "^3.4.13",
+		"prettier": "^3.3.2",
+		"prettier-plugin-svelte": "^3.2.6",
+		"prettier-plugin-tailwindcss": "^0.6.5",
+		"svelte": "^5.0.0",
+		"tailwindcss": "^3.4.9",
 		"vite": "^5.0.3"
 	},
-	"type": "module",
 	"dependencies": {
-		"@sveltejs/adapter-vercel": "^5.4.5",
-		"lucide-svelte": "^0.453.0",
-		"svelte-persisted-store": "^0.11.0"
+		"@sveltejs/adapter-vercel": "^5.4.7",
+		"lucide-svelte": "^0.456.0"
 	}
 }
