@@ -2,10 +2,9 @@
 	import { page } from '$app/stores';
 	import { BookOpen, HouseIcon, Info, UserIcon } from 'lucide-svelte';
 	import NavMarker from './NavMarker.svelte';
-	import { onMount } from 'svelte';
 	import { getUserToken } from '$lib/api';
 
-	let navContainer;
+	let navContainer = $state();
 
 	const navLinks = [
 		{ href: `/`, title: 'Home', icon: HouseIcon },
@@ -19,65 +18,105 @@
 		window.location.href = ``;
 	};
 
-	let loggedIn;
-	onMount(async () => {
+	let loggedIn = $state();
+	$effect(async () => {
 		loggedIn = getUserToken() ? true : false;
 	});
 </script>
 
-<nav class="w-full text-slate-900 bg-white shadow sticky top-0 z-50">
-	<div class="max-w-5xl justify-center items-center gap-4 p-3 mx-auto hidden md:flex">
+<nav class="sticky top-0 z-50 w-full bg-white text-slate-900 shadow">
+	<div class="mx-auto hidden max-w-5xl items-center justify-center gap-4 p-3 md:flex">
 		{#each navLinks as link}
 			{#if link.title !== 'Profile'}
-				<a href={link.href} class="flex px-3 py-2 rounded-lg hover:text-primary-700 duration-300 transition-all {$page.url.pathname === link.href ? 'text-primary-700' : 'text-primary-900'}">
-					<svelte:component this={link.icon} class="mx-auto md:me-2" stroke-width={1.5} />
+				<a
+					href={link.href}
+					class="flex rounded-lg px-3 py-2 transition-all duration-300 hover:text-primary-700 {$page
+						.url.pathname === link.href
+						? 'text-primary-700'
+						: 'text-primary-900'}"
+				>
+					<link.icon class="mx-auto md:me-2" stroke-width={1.5} />
 					<p>{link.title}</p>
 				</a>
 			{/if}
 		{/each}
 		<div class="ml-0 lg:ml-16">
-			<div class="flex items-center duration-300 transition-all">
+			<div class="flex items-center transition-all duration-300">
 				{#if loggedIn === true}
-					<a href="/profile" class="flex py-2 rounded-lg hover:text-primary-700 duration-300 transition-all {$page.url.pathname === `/profile` ? 'text-primary-700' : 'text-slate-900'}">
+					<a
+						href="/profile"
+						class="flex rounded-lg py-2 transition-all duration-300 hover:text-primary-700 {$page
+							.url.pathname === `/profile`
+							? 'text-primary-700'
+							: 'text-slate-900'}"
+					>
 						<UserIcon stroke-width={1.5} class="md:me-2" />
 						<p>Profile</p>
 					</a>
 				{:else}
 					<p class="flex items-center">
 						<UserIcon stroke-width={1.5} class="md:me-2" />
-						<a href={`/sign-in`} class="hover:text-primary-700 duration-300 transition-all {$page.url.pathname === `/sign-in` ? 'text-primary-700' : 'text-slate-900'}">Login</a>
+						<a
+							href={`/sign-in`}
+							class="transition-all duration-300 hover:text-primary-700 {$page.url.pathname ===
+							`/sign-in`
+								? 'text-primary-700'
+								: 'text-slate-900'}">Login</a
+						>
 						/
-						<a href={`/sign-up`} class="hover:text-primary-700 duration-300 transition-all {$page.url.pathname === `/sign-up` ? 'text-primary-700' : 'text-slate-900'}">Register</a>
+						<a
+							href={`/sign-up`}
+							class="transition-all duration-300 hover:text-primary-700 {$page.url.pathname ===
+							`/sign-up`
+								? 'text-primary-700'
+								: 'text-slate-900'}">Register</a
+						>
 					</p>
 				{/if}
 			</div>
 		</div>
 	</div>
 	{#if ['/profile', '/sign-in', '/sign-up'].includes($page.url.pathname) && loggedIn === false}
-		<div class="md:hidden flex justify-end items-center">
-			<a href={`/sign-in`} class="flex px-3 py-3 text-slate-900 font-bold duration-300 transition-all">
+		<div class="flex items-center justify-end md:hidden">
+			<div class="flex px-3 py-3 font-bold text-slate-900 transition-all duration-300">
 				<UserIcon stroke-width={1.5} />
-				<a href={`/sign-in`} class="hover:text-primary-700 duration-300 transition-all {$page.url.pathname === `/sign-in` ? 'text-primary-700' : 'text-slate-900'}">Login</a>
+				<a
+					href={`/sign-in`}
+					class="transition-all duration-300 hover:text-primary-700 {$page.url.pathname ===
+					`/sign-in`
+						? 'text-primary-700'
+						: 'text-slate-900'}">Login</a
+				>
 				/
-				<a href={`/sign-up`} class="hover:text-primary-700 duration-300 transition-all {$page.url.pathname === `/sign-up` ? 'text-primary-700' : 'text-slate-900'}">Register</a>
-			</a>
+				<a
+					href={`/sign-up`}
+					class="transition-all duration-300 hover:text-primary-700 {$page.url.pathname ===
+					`/sign-up`
+						? 'text-primary-700'
+						: 'text-slate-900'}">Register</a
+				>
+			</div>
 		</div>
 	{:else if $page.url.pathname === '/profile' && loggedIn === true}
-		<div class="md:hidden flex justify-end items-center">
-			<button on:click={handleSignOut} class="px-4 py-3 rounded-lg">Sign Out</button>
+		<div class="flex items-center justify-end md:hidden">
+			<button onclick={handleSignOut} class="rounded-lg px-4 py-3">Sign Out</button>
 		</div>
 	{/if}
-	<div bind:this={navContainer} class="md:hidden fixed bottom-0 left-0 z-1 w-full bg-white rounded-t-2xl shadow p-3 flex items-stretch justify-between gap-4 overflow-hidden">
+	<div
+		bind:this={navContainer}
+		class="z-1 fixed bottom-0 left-0 flex w-full items-stretch justify-between gap-4 overflow-hidden rounded-t-2xl bg-white p-3 shadow md:hidden"
+	>
 		<NavMarker {navLinks} {navContainer} />
 		{#each navLinks as link}
 			<a
 				href={link.href}
-				class="nav-link flex flex-col items-center px-2 py-2 rounded-lg hover:text-primary-700 duration-300 transition-all {$page.url.pathname === link.href
+				class="nav-link flex flex-col items-center rounded-lg px-2 py-2 transition-all duration-300 hover:text-primary-700 {$page
+					.url.pathname === link.href
 					? 'active text-primary-700'
 					: 'text-primary-900'}"
 			>
-				<svelte:component this={link.icon} class="mx-auto md:me-2" stroke-width={1.5} />
-				<p class="text-xs whitespace-nowrap">{link.title}</p>
+				<link.icon class="mx-auto md:me-2" stroke-width={1.5} />
+				<p class="whitespace-nowrap text-xs">{link.title}</p>
 			</a>
 		{/each}
 	</div>
