@@ -20,26 +20,33 @@
 	}
 </script>
 
-<section class="flex h-[calc(100vh-8rem)] sm:h-auto sm:mt-8 items-center justify-center">
-	<div class="p-6 rounded-lg shadow-lg max-w-md w-full hover:-translate-y-1 duration-300 transition-all">
-		<div class="flex mb-6 justify-between items-center">
+<section class="flex h-[calc(100vh-8rem)] items-center justify-center sm:mt-8 sm:h-auto">
+	<div
+		class="w-full max-w-md rounded-lg p-6 shadow-lg transition-all duration-300 hover:-translate-y-1"
+	>
+		<div class="mb-6 flex items-center justify-between">
 			<h1 class="text-2xl font-bold text-gray-700">Forgot Password</h1>
 		</div>
 
-		<form on:submit={handleSubmit}>
+		<form onsubmit={handleSubmit}>
 			<div class="mb-4">
 				<label for="email">Email</label>
 				<input
 					name="email"
 					id="email"
 					type="email"
-					class="mt-1 block w-full px-3 py-2 border border-gray-300 rounded-md shadow-sm focus:outline-none focus:ring-primary-500 focus:border-primary-500 sm:text-sm"
+					class="focus:ring-primary-500 focus:border-primary-500 mt-1 block w-full rounded-md border border-gray-300 px-3 py-2 shadow-sm focus:outline-none sm:text-sm"
 					required
 				/>
 			</div>
 
 			<div class="text-center">
-				<button type="submit" class="bg-primary-800 text-white px-8 py-2 rounded-lg hover:bg-primary-700 duration-300 transition-all"> Reset Password </button>
+				<button
+					type="submit"
+					class="bg-primary-800 hover:bg-primary-700 rounded-lg px-8 py-2 text-white transition-all duration-300"
+				>
+					Reset Password
+				</button>
 			</div>
 		</form>
 	</div>
